committing to checkout master
