+++ client/components/RecordButtons.js
@@ -1,0 +1,24 @@
+import React from 'react';
+import { connect } from 'react-redux';
+
+
+const RecordButtons = () => {
+
+
+    return (
+        <div>
+            <h1>Record Buttons Here</h1>
+
+        </div>
+
+    )
+
+}
+
+const mapState = (state) => {
+    return {
+    }
+}
+
+
+export default connect(mapState)(RecordButtons);
