+++ client/components/user-home.js
@@ -1,34 +1,34 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import {connect} from 'react-redux'
+// import React from 'react'
+// import PropTypes from 'prop-types'
+// import {connect} from 'react-redux'
 
-/**
- * COMPONENT
- */
-export const UserHome = (props) => {
-  const {email} = props
+// /**
+//  * COMPONENT
+//  */
+// export const UserHome = (props) => {
+//   const {email} = props
 
-  return (
-    <div>
-      <h3>Welcome, {email}</h3>
-    </div>
-  )
-}
+//   return (
+//     <div>
+//       <h3>Welcome, {email}</h3>
+//     </div>
+//   )
+// }
 
-/**
- * CONTAINER
- */
-const mapState = (state) => {
-  return {
-    email: state.user.email
-  }
-}
+// /**
+//  * CONTAINER
+//  */
+// const mapState = (state) => {
+//   return {
+//     email: state.user.email
+//   }
+// }
 
-export default connect(mapState)(UserHome)
+// export default connect(mapState)(UserHome)
 
-/**
- * PROP TYPES
- */
-UserHome.propTypes = {
-  email: PropTypes.string
-}
+// /**
+//  * PROP TYPES
+//  */
+// UserHome.propTypes = {
+//   email: PropTypes.string
+// }
