+++ public/index.html
@@ -3,16 +3,12 @@
   <head>
     <meta charset='utf-8'>
     <meta name='viewport' content='width=device-width initial-scale=1.0'>
-    <title>Boilermaker</title>
+    <title>Power Prose</title>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta.2/css/bootstrap.min.css" integrity="sha384-PsH8R72JQ3SOdhVi3uxftmaW6Vc51MKb0q5P2rRUpPvrszuE4W1povHYgTpBfshb" crossorigin="anonymous">
     <link rel="stylesheet" href="/style.css" />
     <script defer src="/bundle.js"></script>
   </head>
   <body>
-    <div class="on-button-container">
-      <button class="on-button">START MY ROUTINE</button>
-    </div>
-    <div>
-      <button class="stop-button">Stop</button>
-    </div>
+    <div id="app"></div>
   </body>
 </html>
